+++ test/components/breadcrumbs_test.rb
@@ -13,7 +13,7 @@ class PrimerBreadcrumbsTest < Minitest::Test
 
   def test_renders_single_item
     render_inline(Primer::Beta::Breadcrumbs.new) do |component|
-      component.item { "Home" }
+      component.item(href: "/") { "Home" }
     end
 
     assert_selector("nav[aria-label='Breadcrumb'] .breadcrumb-item", text: "Home")
@@ -21,9 +21,9 @@ class PrimerBreadcrumbsTest < Minitest::Test
 
   def test_renders_multiple_items
     render_inline(Primer::Beta::Breadcrumbs.new) do |component|
-      component.item { "Home" }
-      component.item { "About" }
-      component.item { "Team" }
+      component.item(href: "/") { "Home" }
+      component.item(href: "/about") { "About" }
+      component.item(href: "/about/team") { "Team" }
     end
 
     assert_selector("nav[aria-label='Breadcrumb'] .breadcrumb-item", count: 3)
@@ -32,20 +32,20 @@ class PrimerBreadcrumbsTest < Minitest::Test
   def test_renders_links_when_specified
     render_inline(Primer::Beta::Breadcrumbs.new) do |component|
       component.item(href: "/") { "Home" }
-      component.item { "About" }
+      component.item(href: "/about") { "About" }
     end
 
-    assert_selector("nav[aria-label='Breadcrumb'] .breadcrumb-item a", count: 1)
+    assert_selector("nav[aria-label='Breadcrumb'] .breadcrumb-item a", count: 2)
     assert_selector("nav[aria-label='Breadcrumb'] .breadcrumb-item a[href='/']")
   end
 
-  def test_does_not_render_a_link_when_item_is_selected
+  def test_automatically_selects_last_item
     render_inline(Primer::Beta::Breadcrumbs.new) do |component|
-      component.item(href: "/", selected: true) { "Home" }
-      component.item { "About" }
+      component.item(href: "/") { "Home" }
+      component.item(href: "/about") { "About" }
     end
 
-    refute_selector("nav[aria-label='Breadcrumb'] .breadcrumb-item a")
+    assert_selector("li a[aria-current='page'].breadcrumb-item-selected", text: "About")
   end
 
   def test_status